Refactor a little
